refactor: extract domain actions from App
